refactor(voicecli): create ui/ sub-package (T5 of #83)
8 modules moved under voicecli.ui (ui_sounds.py → sounds.py rename, drops
redundant prefix). No behavior change.
