+++ pyrightconfig.json
@@ -8,9 +8,9 @@
     "src/voicecli/engines"
   ],
   "ignore": [
-    "src/voicecli/overlay.py",
-    "src/voicecli/overlay_draw.py",
-    "src/voicecli/overlay_audio.py",
+    "src/voicecli/ui/overlay.py",
+    "src/voicecli/ui/overlay_draw.py",
+    "src/voicecli/ui/overlay_audio.py",
     "src/voicecli/nats",
     "src/voicecli/adapters/nats",
     "src/voicecli/runtime/daemon.py",
