+++ src/voicecli/ui/__init__.py
@@ -1,0 +1,1 @@
+"""voicecli.ui — UI-layer modules (overlay, clipboard, sounds, STT client, NATS recorder, Telegram)."""
