+++ src/voicecli/ui/overlay_draw.py
