+++ src/voicecli/ui/stt_client.py
